+++ drizzle/0001_spotty_salo.sql
@@ -1,0 +1,4 @@
+ALTER TABLE "articles" ADD COLUMN "excerpt" text;--> statement-breakpoint
+ALTER TABLE "articles" ADD COLUMN "meta_title" text;--> statement-breakpoint
+ALTER TABLE "articles" ADD COLUMN "meta_description" text;--> statement-breakpoint
+ALTER TABLE "articles" ADD COLUMN "og_image" text;